+++ app.log
@@ -107,3 +107,29 @@
 2025-03-16 18:11:32,075 - INFO - Пользователь test_test успешно вошел в систему.
 2025-03-16 18:12:03,665 - INFO - Пользователь test_test успешно вошел в систему.
 2025-03-16 18:12:06,778 - WARNING - Отрицательная цена билета.
+2025-03-17 10:42:55,992 - WARNING - Пользователь с логином Jumomi или почтой уже существует.
+2025-03-17 10:47:14,778 - WARNING - Пользователь с логином Jumomi или почтой уже существует.
+2025-03-17 10:47:48,440 - INFO - Пользователь test_test успешно вошел в систему.
+2025-03-17 10:47:53,187 - WARNING - Отрицательная цена билета.
+2025-03-17 10:48:08,130 - INFO - Пользователь test_test успешно вошел в систему.
+2025-03-17 10:48:16,331 - WARNING - Пользователь с ID 7 не является организатором.
+2025-03-24 09:30:59,769 - WARNING - Неверный пароль для пользователя Jumomi.
+2025-03-24 09:31:54,274 - WARNING - Мероприятие с ID 999 не найдено.
+2025-03-24 09:31:54,287 - WARNING - Пользователь с ID 999 не найден.
+2025-03-24 09:31:54,306 - INFO - Пользователь с ID 1 успешно приобрел билет на мероприятие ID 1.
+2025-03-24 09:33:20,859 - WARNING - Попытка входа с несуществующим логином: test_user.
+2025-03-24 09:33:33,507 - WARNING - Попытка входа с несуществующим логином: JUMOMI.
+2025-03-24 09:34:00,137 - WARNING - Пользователь с логином test или почтой уже существует.
+2025-03-24 09:34:13,195 - WARNING - Неверный пароль для пользователя test.
+2025-03-24 09:34:25,408 - WARNING - Неверный пароль для пользователя test.
+2025-03-24 09:35:04,393 - INFO - Пользователь test_user успешно зарегистрирован.
+2025-03-24 09:35:10,598 - WARNING - Пользователь с ID None не является организатором.
+2025-03-24 09:35:24,563 - INFO - Пользователь test_user успешно вошел в систему.
+2025-03-24 09:37:00,855 - INFO - Пользователь test_user успешно вошел в систему.
+2025-03-24 09:37:06,015 - WARNING - Отрицательная цена билета.
+2025-03-24 09:44:02,075 - INFO - Пользователь test_user успешно вошел в систему.
+2025-03-24 09:44:08,171 - WARNING - Отрицательная цена билета.
+2025-03-24 09:45:12,205 - WARNING - Мероприятие с ID 999 не найдено.
+2025-03-24 09:45:12,217 - WARNING - Пользователь с ID 999 не найден.
+2025-03-24 09:45:12,236 - INFO - Пользователь с ID 1 успешно приобрел билет на мероприятие ID 1.
+2025-03-24 09:46:20,399 - WARNING - Неверный пароль для пользователя Jumomi.
